ScheduleDisplay component
Schedules are now also in their own component, rather than being rendered directly on the page.
